\name{isfs-package}
\alias{isfs-package}
\alias{isfs}
\docType{package}
\title{
Support for NCAR EOL ISFS, Integrated Surface Flux System
}
\description{
This package provides R functions to access and manipulate time series from the NCAR EOL Integrated Surface Flux System.
}
\details{
\tabular{ll}{
Package: \tab isfs\cr
License: \tab BSD_2_clause + file LICENSE\cr
URL: \tab http://github.com/ncareol/R-eol.git\cr
Authors@R: \tab person("Gordon","Maclean",role=c("aut","cre"),
            email="maclean@ucar.edu")\cr
Depends: \tab R (>= 3.0), methods, eolts\cr
Imports: \tab quantreg\cr
}
}
\author{
Gordon Maclean
}
\references{
}
\keyword{ package ts }
\seealso{
\code{\link[eolts:eolts-package]{eolts}}
\code{\link{plot_sunangle}}
\code{\link{plot_sunshadow}}
}
\examples{
}
