\name{isfs-package}
\alias{isfs-package}
\alias{isfs}
\docType{package}
\title{
Support for NCAR EOL ISFS, Integrated Surface Flux System
}
\description{
This package provides R functions to access and manipulate time series from the NCAR EOL Integrated Surface Flux System.
}
\details{
\tabular{ll}{
Package: \tab isfs\cr
License: \tab BSD_2_clause + file LICENSE\cr
URL: \tab http://github.com/ncareol/R-eol.git\cr
Authors@R: \tab person("Gordon","Maclean",role=c("aut","cre"),
            email="maclean@ucar.edu")\cr
Depends: \tab R (>= 3.0), methods, eolts\cr
Imports: \tab quantreg\cr
}
}
\author{
Gordon Maclean
}
\references{
}
\keyword{ package }
\keyword{ ts }
\seealso{
\tabular{ll}{
?command \tab link\cr
\code{package? eolts} \tab \code{\link[eolts:eolts-package]{eolts}}\cr
\code{class? dat} \tab \code{\linkS4class{dat}}\cr
\code{? dat} \tab \code{\link{dat}}\cr
\code{class? prep} \tab \code{\linkS4class{prep}}\cr
\code{? prep} \tab \code{\link{prep}}\cr
\code{? plot_sunangle} \tab \code{\link{plot_sunangle}}\cr
\code{? plot_sunshadow} \tab \code{\link{plot_sunshadow}}\cr
}
}
\examples{
}
